Table entity modified.

+++ Maven szerver/Maven_Rendeles/src/main/java/org/example/Models/TableEntity.java
@@ -32,7 +32,7 @@ public class TableEntity {
     @Schema(description = "Székek száma az asztalnál", example = "3")
     @Min(value = 2, message = "error.table.capacity.min")
     @NotNull(message = "error.table.capacity.notset")
-    private byte capacity;
+    private int capacity;
 
 
 }
